Update Python_CodeName_Game.py
Fixed Mole feature by adding global variable "gui" to both New_Game() and task() methods.  An instance named "gui"  of class "Board" is created in New_Game()  and used in task() to call the Mole_Agent method.  Adjusted constant of 10.03 to 10.5 at the end of task() to trigger the insertion of a mole roughly once per minute with frequency setting of 5.  Settings less than five will subsequently reduce the rate of appearance of moles proportionately.